refactor: deduplicate blog posts fetching
