+++ src/routes/writing/[tag].json.ts
@@ -1,7 +1,7 @@
 import type { RequestHandler } from '@sveltejs/kit';
 import type { Locals, WritingPreview } from '$lib/types';
-import { basename } from 'path';
 import slugify from 'slugify';
+import { getBlogPosts } from './_api';
 
 interface GetResult {
 	prettyTagName?: string;
@@ -9,24 +9,12 @@ interface GetResult {
 }
 
 export const get: RequestHandler<Locals, GetResult> = async ({ params: { tag } }) => {
-	const svxFiles = import.meta.globEager('./*.svx');
-
-	const importedFiles: WritingPreview[] = Object.entries(svxFiles).map(([path, { metadata }]) => {
-		const { title, description, datetime, tags } = metadata as Omit<WritingPreview, 'slug'>;
-
-		return {
-			title,
-			description,
-			datetime,
-			tags,
-			slug: basename(path, '.svx')
-		};
-	});
+	const blogPosts = getBlogPosts();
 
 	const articles: WritingPreview[] = [];
 	let prettyTagName: string | undefined;
 
-	for (const blogPost of importedFiles) {
+	for (const blogPost of blogPosts) {
 		for (const blogPostTag of blogPost.tags) {
 			const sluggifiedTag = slugify(blogPostTag, {
 				lower: true
